basics/10-channel-diarization/requirements.txt
# Speechmatics RT SDK (multi-channel diarization)
speechmatics-rt>=0.5.0

# Environment management
python-dotenv>=1.2.1
